+++ react-app/src/App.js
@@ -6,6 +6,7 @@ import NavBar from "./components/Navbar";
 import ProtectedRoute from "./components/auth/ProtectedRoute";
 import UsersList from "./components/UsersList";
 import User from "./components/User";
+import HomePage from "./components/HomePage";
 import { sessionAuthenticate } from "./store/session";
 import { useSelector, useDispatch } from "react-redux";
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
@@ -42,44 +43,42 @@ function App() {
               isOpen={isOpen} setIsOpen={setIsOpen} />
       <div className={`content${isOpen?" open":""}`}>
       <TabBar/>
-      <Switch>
-        <Route path="/login" exact={true}>
-          <LoginForm
+      <div className="page-container">
+        <Switch>
+          <Route path="/login" exact={true}>
+            <LoginForm
+                authenticated={authenticated}
+                setAuthenticated={setAuthenticated}
+              />
+            </Route>
+            <Route path="/sign-up" exact={true}>
+              <SignUpForm
+                authenticated={authenticated}
+                setAuthenticated={setAuthenticated}
+              />
+            </Route>
+            <ProtectedRoute
+              path="/users"
+              exact={true}
               authenticated={authenticated}
-              setAuthenticated={setAuthenticated}
-            />
-          </Route>
-          <Route path="/sign-up" exact={true}>
-            <SignUpForm
+            >
+              <UsersList />
+            </ProtectedRoute>
+            <ProtectedRoute
+              path="/users/:userId"
+              exact={true}
               authenticated={authenticated}
-              setAuthenticated={setAuthenticated}
-            />
-          </Route>
-          <ProtectedRoute
-            path="/users"
-            exact={true}
-            authenticated={authenticated}
-          >
-            <UsersList />
-          </ProtectedRoute>
-          <ProtectedRoute
-            path="/users/:userId"
-            exact={true}
-            authenticated={authenticated}
-          >
-            <User />
-          </ProtectedRoute>
-          <Route path="/" exact={true}>
-            <h1>My Home Page</h1>
-          </Route>
-          <Route path="/edit-profile" exact={true} authenticated={authenticated}>
-            <h1>Hi</h1>
-          </Route>
-          <Route path="/" exact={true}>
-            <a href="/login" className="login">Login</a>
-            <a href="/sign-up" className="signup">Signup</a>
-          </Route>
-        </Switch>
+            >
+              <User />
+            </ProtectedRoute>
+            <Route path="/" exact={true}>
+              <HomePage />
+            </Route>
+            <Route path="/edit-profile" exact={true} authenticated={authenticated}>
+              <h1>Hi</h1>
+            </Route>
+          </Switch>
+        </div>
       </div>
     </BrowserRouter>
   );
