+++ react-app/src/components/auth/LoginForm.js
@@ -48,43 +48,43 @@ const LoginForm = ({ authenticated, setAuthenticated }) => {
   }
 
   return (
-    <div className="page-container">
-      <form onSubmit={onLogin} className="loginform">
-          <div>
-            <label htmlFor="email">email</label>
-            <span className={email === ""? " " : "quoted"}>
-            <input
-              name="email"
-              type="text"
-              placeholder="null"
-              value={email}
-              onChange={updateEmail}
-              size={email.length-1}
-            />
-            </span>
-          </div>
-          <div>
-            <label htmlFor="password">password</label>
-            <span className={password === ""? " " : "quoted"}>
-            <input
-              name="password"
-              type="password"
-              placeholder="null"
-              value={password}
-              onChange={updatePassword}
-              size={password.length}
-            />
-            </span>
-          </div>
-          <div><span style={{color:"#dcb862"}}>login</span>(<span style={{color:"#2ba2ff"}}>email</span>, <span style={{color:"#2ba2ff"}}>password</span>);</div>
-          <button type="submit">{`> `}node login.js</button>
-      </form>
-      <div className="errorsLog" >
-        {errors.map((error) => (
-          <div>{error}</div>
-        ))}
-      </div>
-    </div>
+      <>
+        <form onSubmit={onLogin} className="loginform">
+            <div>
+              <label htmlFor="email">email</label>
+              <span className={email === ""? " " : "quoted"}>
+              <input
+                name="email"
+                type="text"
+                placeholder="null"
+                value={email}
+                onChange={updateEmail}
+                size={email.length-1}
+              />
+              </span>
+            </div>
+            <div>
+              <label htmlFor="password">password</label>
+              <span className={password === ""? " " : "quoted"}>
+              <input
+                name="password"
+                type="password"
+                placeholder="null"
+                value={password}
+                onChange={updatePassword}
+                size={password.length}
+              />
+              </span>
+            </div>
+            <div><span style={{color:"#dcb862"}}>login</span>(<span style={{color:"#2ba2ff"}}>email</span>, <span style={{color:"#2ba2ff"}}>password</span>);</div>
+            <button type="submit">{`> `}node login.js</button>
+        </form>
+        <div className="errorsLog" >
+          {errors.map((error) => (
+            <div>{error}</div>
+          ))}
+        </div>
+      </>
   );
 };
